requirements.txt
playwright
python-dotenv
pandas
ffmpeg-python